+++ packages/go_router/lib/go_router.dart
@@ -19,7 +19,6 @@ export 'src/configuration.dart'
 export 'src/misc/extensions.dart';
 export 'src/misc/inherited_router.dart';
 export 'src/pages/custom_transition_page.dart';
-export 'src/platform.dart' show UrlPathStrategy;
 export 'src/route_data.dart' show GoRouteData, TypedGoRoute;
 export 'src/router.dart';
 export 'src/typedefs.dart'