+++ packages/pigeon/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 4.2.3
+
+* [java] Adds assert `args != null`.
+* [java] Changes the args of a single element to `ArrayList` from `Arrays.asList` to `Collections.singletonList`.
+* [java] Removes cast for `Object`.
+
 ## 4.2.2
 
 * Removes unneeded custom codecs for all languages.